Extract headland logic to MainViewModel.Headland.cs (886 lines)
Moved from MainViewModel.cs (5,643 -> 4,795 lines) to dedicated
partial class file:
- ComputeSegmentOffset (Clipper2 + edge-based offset)
- BuildHeadlandFromSegments (chain merge, closed loop, polygon split)
- RemoveSelfIntersections
- FindLineHeadlandIntersection + Farthest variant
- BuildOffsetLineWithExtensions
- LineLineIntersection, LineSegmentIntersection, IsPointInPolygon
- CalculateSignedArea
- SaveHeadlandSegments

No functional changes - all 67 tests pass.
